Storing tree elements in a linked-list
Previously we held the elements in a slice. This works fine, but has the
draw back of containing pointers, which we wish to completely eliminate
from the tree data-structure itself.

Instead of a slice we now build the elements into a linked-list. The
linked list is built from elements taken from an ObjectStore. So no Go
pointers are needed, just ObjectPointers.