Adding a package for reading LDS parcel data
This data describes all of the land parcels in New Zealand. In this new
package we provide a convenient struct representation for this data.

The fact that land parcels are not point-like but are complex polygons
of long-lat points indicates that the point-like quadtree isn't the
perfect data-structure for this data.

It's not clear how to find a shape based data-structure for this though.
Likely we will just push all the points of the polygon into the
quad-tree and maybe require an enlarged search area to best-effort deal
with the case where your search area falls entirely inside a polygon.

Maybe there is some useful insight that provide a clean approach to
dealing with this.