Separated LineNum and Error from Parcel data
Previously the struct was called RawCSVEntry, which is a completely
meaningless name. This struct had meta-data, line number and error mixed
in with land parcel data.

Now we have ParcelData struct, which has the actual land parcel data.
This is wrapped in a CSVParcelData struct. This struct has the line
number and error meta-data along side the parcel data. This makes it
clearer when we find an error in a csv line and we write the line number
and the error data, but leave the parcel data zeroed.