Changed View coordinate system
Previously we said that the view's coordinates conceptually started a
0,0 as the top left point of the positive coordinates.

This means that
the left most x is lower than the right most x, which matches the
standard way of drawing a graph from highschool. Which is fine.

But it also means that the top y coordinate is less than the bottom y
coordinate i.e. the positive y coordinates start with zero at the top
and grow downwards. This is the opposite of how you would draw a graph
in highschool. It also matches badly with convenient mental models for
latitude values, where higher values are north (up) and lower values are
south (down).

The reason the previous coordinate system was chosen was because some
graphics coordinate systems start with the top left corner of the screen
being 0,0 and grow rightwards and downwards. This is fine for graphics,
but we are specifically representing long/lat values here.