Objectstore.Free(ObjectPointer) method added
This method allows us to 'free' an object, this object can then be
reused for future Get() calls.

To achieve this we added a very simple linked list wrapper to each
object.

The ObjectStore itself has a pointer to a free object (this pointer may
be nil). This pointer is called the root free pointer, all freed objects
are reachable starting from the root free pointer and following the
linked list.

When allocating a new object we first check to see if a
previously freed object is available. If a freed object is available we
return that and then set the root free pointer to point at the next
object in the linked list (which might be nil).

We maintain an important property. Any freed object (which has not been
re-allocated) has a non-nil pointer to another freed object. In the case
where there isn't another freed object to point to, the freed object
points to itself.

This allows us to test whether an object being accessed via Get is
actually free. We can also test if an object being Free(ed) is already
free. In both of these cases the method calls will panic.

After writing all of this code, we seem to have a mixture of terminology
developing. We Get() a new object and then increment a field called
allocs. This starts to feel confusing. We should try to tidy up the
naming of these fields and methods in the future.