Separate meta-data from objects in ObjectStore
This is _probably_ very slightly slower than storing the metadata right
next to the objects. We make this change because it is a simpler
implementation approach for the ByteStore. I would prefer to have both
stores use the same implementation approach to start with. We can make
specific optimisations later, if needed.