ByteStore now allocates from pre-sized byte slabs
These slabs are all power of 2 sized, although the first 4 bytes is used
to record the size of the actual allocation. So for example, if you want
to allocate 4 bytes, you will need at least an 8 byte sized slab to
allow for the size to be written.

This approach is adopted to make it easier to free allocated byte
slices. For this to be reasonably simple we would rather allocate and
free from pre-sized byte slices.

The smallest slab used is 8 bytes, so 0, 1, 2, 3 and 4 byte allocations
all use this smallest slab. The remaining slabs are simply successive
powers of 2 in size, e.g. 8, 16, 32, 64, ...

The hardest part of implementing this was building a reliable mapping
from allocation size to an index starting at 0 (so we could lookup a
slab by index using the desired allocation size). Some subtle code was
developed to achieve this, there may be a simpler way.