Extended allocator slabs to cover all of uint32
This means that you can now allocate a slice of any size inside of
uint32. Take care as allocating a single very large slice will create
1024 of those slots when allocating a whole chunk. If your allocation is
1 Gig we will now allocate 1024 Gigs for you. This needs to be fixed.

We created a separate function which initialises the slice of slabs. The
slabs are now organised as follows:

0: The specialty 0 sized slab allocator
1: 2^0 sized slab allocator
2: 2^1 sized slab allocator
3: 2^2 sized slab allocator
4: 2^3 sized slab allocator
...
31: 2^30 sized slab allocator
32: 2^31 sized slab allocator
33: 2^32 - 1 sized slab allocator

The final allocator looks a bit odd, because you might expect it to be
2^32 sized, but 2^32 is 0 in a uint32, so we just take max-value here
instead.

We removed the existing tests, which tested the lower level indexForSize
functions and now we just test that the completely initialised slab
slice is properly set up.

One observation that I made while writing and debugging the tests is
that evenly distributed random numbers in uint32 are almost all very
large numbers. This surprised me at first, but after some thinking I
realised that almost half the numbers in a uint32 are larger than 2^31.
Furthermore a solid majority of these numbers are larger than 2^30 and
so on. There are just more large numbers than small ones if we group
them into powers of two (like our allocator does). So we did some extra
processing to ensure that our tests generate an even spread of large and
small numbers across the powers of two groups that we care about.