Add panics if an alloc() call is made with a bad size
We protect against allocs with sizes which are too large, these just
can't be satisfied by a slab with a too small slot size.

We protect against allocs with sizes which are less than half of the
slot size. These allocs can be satisfied by the slab, but is very space
inefficient. This indicates an unrecoverable programming error in the
ByteStore.