Adding linkedlist package
This package builds on objectstore to make an easy-to-use linked list
implementation. The linked list is an invasive linked list, which binds
the list and the data together very tightly. For example a data item
cannot be in two lists at once, as you might naturally do with two
slices.