Linked lists are empty when created
Previously an linked list was created with a single element. This is
very unusual behaviour and would be surprising to anyone. An empty list
is the obvious initial state for a created list.

This change simplified the tests and made them look more natural.

This also has the advantage of making nil list pointers, the
representation of an empty list, more common. Previously the empty list
case was under tested because you had to specifically empty a list in
order to test the empty cases. We still need to add more explicit empty
list tests, but this is a good step.