Adding Push/Peak/Pop methods to linkedlist.List
These methods all come in a Head and Tail version. The Head version
always operates on the first position of the list, while the Tail
version operates on the last position of the list.

So the full list is:

PushHead
PushTail
PeakHead
PeakTail
PopHead
PopTail

There are no tests written for these methods yet. This will come next.