Fixing an object recycle bug in objectstore
This bug is caused by the fact that we require that allocated-then-freed
objects have a non-nil nextFree pointer in their meta data.

This is helpful because it means we can identify free pointers and panic
if client code tries to use them. The last free pointer points to
itself.

The problem comes when we try to allocate the last free pointer. Because
its 'nextFree' pointer points to itself, we may just set the rootFree
pointer to be the pointer we just allocated. We avoid this with a simple
if statement, and we've added a test to cover this problem.

There is a higher level problem there though. This bug existed and was
fixed in the bytestore package (which uses the same metadata approach to
managing freed objects). It's definitely a shame that we _forgot_ to
check if this bug existed in the other package. But I think the larger
problem is that the unit tests we have right now don't catch these bugs.
Looking over the unit tests they are carefully written and well
commented, they look very well thought out. But I worry that the current
approach makes the tests too hard to write, and we miss deeper tests of
more complex scenarios.

We can look to the table-driven tests of the linkedlist package (these
tests actually found the objectstore bug). We should consider writing
these kinds of tests for the objectstore and bytestore packages.