First complete fuzz test of object_store
Not a bad basic start. likely the fuzz_util needs unit testing as it
looks complex enough to contain many bugs.

We learned an interesting property of the `Free` method. Currently it
makes a best effort to detect if you are freeing an already free object.
If it detects this it panics. However this behaviour is not guaranteed.
Because it's possible that a freed object may be re-allocated by another
Alloc call. In this case instead of panicking, the Free call will
quietly free someone else's object. So we can't reliably test the double
Free panic in these tests as we can't guarantee they won't be
re-allocated.

This observation really suggests that it might be worth adding an extra
field to a pointer, a _generation_ counter or similar, which
differentiates these cases and panics reliably. At least at first double
frees and related behaviour are likely bugs which will be hard to track
down.

There will be a performance cost, but it's likely worth it. Time will
tell.