Adding fuzz testing for byte_store
This is a copy/pasted version of the objectstore fuzz testing. The end
result is that the tests are very similar between the two of them. We
should refactor these to have a reusable test package for fuzzing.

One thing that should be noted is that there is a recurring challenge of
generating a range of allocation sizes. Evenly distributed random
numbers have a very strong bias towards generating large numbers. We
have a solution to it here, but we should look closer at this and test
the distribution of size classes generated by our current approach.