Added fuzzutil package
This is a first step toward a set of utilities to make it easy to fuzz
components which don't work directly on basic byte or string data.

We introduce two main component.

1: fuzzutil.TestRun

A TestRun is built of a slice of fuzzutil.Step objects. There is a
generic constructor where client code can supply a function which builds
test steps using fuzzing []byte data.

2: fuzzutil.ByteConsumer

A ByteConsumer wraps the basic []byte that was provided to us by the
fuzzer. It converts parts of the []byte data into useful data, like
[]byte or integer values of different sizes. As each piece of data is
retrieved the internal []byte are used up.

The fuzz tests for objectstore and bytestore fuzz tests are moved
directly into the packages they test.