example restriction for Independent Allocs
An example of an implementation restriction for the 'Independent Allocs'
section in the object_store top level comments was added.

As I write this I realised that the current implementation _requires_
some kind of concurrency protection to allow for safe concurrent calls
to Get() on some set of safely published References, while independent
calls are made to Alloc() on another goroutine.

This is because calls to Alloc() may need to append a new slice to the
objects field (objects [][]O) in the objectstore.Store. This is
definitely a write to a piece of data which must be read by every call
to Get().

Unless this approach is totally re-thought and re-implemented we will
need to add at least this amount of concurrency control to the
objectstore, the bytestore will need a similar modification.

This is probably a natural place to stop writing more documentation on
objectstore and actually fix the (quite large) hole we just found. We
can come back to specifying the behaviour of the objectstore after this
is fixed and see if we can find any other problems.