Replacing unsafeutil package with funsafe
The unsafeutil package is now folded into the larger flib module. The
flib module is preferred as it's a single module containing all of the
helpful packages for doing performance oriented things.

This package is in v0.0.1 because it currently contains some concurrent
queue implementations which really need to be extracted out and moved
into a separate dedicated module of their own.