Remove meta lookups
Previously we kept a [][]meta which contained a reference to the next
free object in the Store. By convention if the next-free reference was
nil then the object was not-free and vise-versa.

We have moved the next-free reference into the main object table, which
now looks like [][]object[O]. This simplifies the lookups as the is-free
meta-data sits right next to the object itself.

Most importantly this will make concurrent read/writes much easier in
the future. Trying to concurrently update both the object and the
meta-data in two different lookup tables will be very difficult. Making
it one location to manage is a dramatic simplification.