Allocate chunks using mmap
If we allocate the chunks using mmap then the objects will sit outside
the view of the garbage collector.

When we allocate the objects using Go's allocator the backing data of
the slice is subject to being moved - if Go ever manages to build a
moving garbage collector.

The biggest advantage of this is that we can then de-reference an
allocated object using a uintptr. Using a uintptr to hold onto a
reference to an object is _usually_ genuinely unsafe, as in won't work
in practice don't-do-it. This is for two reasons.

1: The garbage collector doesn't pay attention to uintptr values when
   determining if an object is live or not. So if your uintptr is the
only reference to an object it will get garbage collected at some point.
Then your uintptr points to random memory and not the object you had
hoped for.

2: The garbage collector has the right to move objects after they have
   been allocated. In practice it's usually generational compacting
garbage collectors which do this. It's a great feature in a garbage
collector and Go may do this in the future. Proper pointers (including
unsafe.Pointer) will get updated to point to the new object location if
they are moved, uintptr will not. So again what was pointing to a valid
object is now pointing to where the object used to be, i.e. random
memory.

In our case the garbage collector never collects our objects - they are
manually managed. If we can also prevent these objects from being moved
then our uintptr _should_ remain safe and valid for accessing objects
directly.

If we can access an object directly, instead of following the
chunk/offset lookup as done currently, we have much greater flexibility
in how we organise the object and their chunks etc. This flexibility
will be a great benefit as we move to a concurrency safe Store
implementation.