Reference.GetValue() added
This allows us to de-reference the value held by a Reference without
going through a *Store. Now we can just hold onto a Reference on its own
and not need to pass around a *Store to get to the data.

This is more convenient for users of this package.

More importantly we have eliminated the table lookup process for getting
data from a Reference. This allows us to restructure the way we organise
the data-chunks for allocating data. In particular we need to eliminate
the use of resizing slices to allow concurrency safety to be added to
this package.