read/write lock protects objectstore.Store.objects
This allows different goroutines to safely read/write the objects field.
There are two ways in which the objects field is accessed.

1: When an allocation is being made to a slab which already exists only
   the read lock is required. This is because we don't actually need to
modify any data in objects to acquire an object. The index which
uniquely identifies an allocation is atomically acquired from allocIdx.

2: When an allocation is being made and there isn't an existing slab
   then we need to allocate the new slab _and_ append it to the objects
slice. This requires a write lock to perform safely.

There's a slightly odd looking feature found in the grownObjects. We
map a new slab _before_ we acquire the write lock, and then append it
unconditionally. This could be perceived as incorrect, because there is
a real possibility that another goroutine got the write lock first and
grew the slice for us. So there is a chance that this could
over-allocate more slabs than are needed.

The benefit of this approach is that the most common case is that only a
single slab is needed, and we can mmap it without taking the write lock.
This means that the locked section only needs to append the new slab and
then likely return immediately.

The downside is possible over allocation of slabs. I don't think this is
a bad downside because if the over-allocated slab is needed in the
future then it wasn't wasted. If it is never used then we pay a very low
cost of it, until the memory is accessed it won't actually truly be
allocated and remains a virtual mapping only.

For this I am only considering Linux memory allocation behaviour. If you
are running this on Windows your experience may be different.