lock protects objectstore.Store.rootFree
This allows us to perform the complicated updates on the rootFree safely
with multiple concurrent goroutines.

There's likely a more sophisticated approach available. But it would be
much more complex than the straight ahead mutex we have here. Since
allocating a previously free object is very fast it's likely not a
problem in practice.