Updated objectstore's docs on memory model
Previously we described a system which supported single
constructor/multiple reader, and single writer/multiple readers for safe
concurrent access.

However, because I really want to be able to build a string interning
library with these packages we need to be able to allocate and retrieve
data from multiple goroutines safely. The work to achieve this is done.

Now the documentation must reflect this.