Reflectively check generic type objectstore.Store
We will now panic if a call is made to objectstore.New[O]() with a
generic type which contains _any_ pointers.

The implementation is very barebones. The checking code errors out on
the first pointerful type it finds. It would be better to search for all
of the bad types, and present a quality description of all of the bad
types found.