Encode generation in Reference.Address
The top byte of Reference.Address is reserved for the generation of each
allocation. An allocation increments its generation each time it is
allocated - start at 0 for the first allocation.

The generation is also written into the object's meta-data.

This allows us to best-effort detect when a Reference is being used to
either access or free data which it no longer owns.

This solution is a best-effort and does not detect _all_ bad Reference
access/free operations. This is because the generation metadata is only
a single byte so repeated alloc/free operations on the same object slot
will cause the generation to wrap around. However, it will likely catch
many (maybe most) programming errors early in the development cycle.

We could make the generation system more robust by turning it into a
uint32 or uint64 value, this would catch more/all bad access/free
operations but would mean each Reference would have to double in size.
This is ok for now.