Adding pointerstore package
This package manages untyped fixed sized allocations.

A pointerstore.Store is constructed with two integer arguments, the size
of the allocations (in bytes) and the number of allocations per slab.
These arguments are then modified in the following way

1: The size of the metadata is added to the allocation size. Each
   allocation is large enough to fit the requested allocation size _and_
the metadata which is prepended to every allocation.
2: The new allocation size is then rounded up to the nearest power of
   two.
3: The number of allocations is rounded up to the nearest power of two.

This gives us a total byte allocation which is also a power of two. But
larger than would be expected looking only at the arguments given when
constructing the pointerstore.Store.

The Reference type can provide a uintptr to the data part of the
allocation or the Metadata directly. The generation/nil checking when
accessing data is handled in the Reference.

This package should be directly usable by objectstore.Store. All that is
needed for the objectstore package is to unsafe cast the uintptr
provided by this pointerstore.Reference into the type being managed by
the objectstore.Store.