Adding pointerstore.AllocationConfig type
This type is passed into the constructor pointerstore.New(...) and
determines the size of each slab and the number of objects held in each
slab.

Constructing an AllocationConfig can be done by providing the
size of the objects you want to allocate and then _either_ the number of
objects per slab or the size of each slab.

The AllocationConfig adjusts the numbers provided to it, to allow for
object meta-data as well as ensuring that all numbers are power of two.

All numbers are adjusted to a nearby power-of-two value to ensure that
allocations fit neatly along cache/page lines.

The allocation config is then made available by pointerstore.Store. This
allows users to understand exactly what is being allocated and into
what.