objectstore.Store can allocate multiple types
Previously each objectstore.Store managed a single type. This meant that
if you were building a datastructure composed of multiple types you
would need to manage multiple Store instances. Now you only need a
single store for all of your types.

This moves validating the type being allocated into the allocation
method itself. Previously this was done when the Store was constructed,
now it must be done with each allocation. Currently we perform the
reflective check for each allocation, but in the future we will cache
the allocated types for faster repeat validation.

Go does not allow for parameterised methods. So we use a parameterised
function, which looks like

Alloc[T any](*Store) Reference[T]

This looks awkward and is not a _natural_ feeling feature. But right now
there doesn't seem to be an alternative. And it works perfectly fine in
practice.

Similarly we have the same function for freeing allocations.

Free[T any](*Store, Reference[T])

Same reason, same discussion as above. Feels awkward - works fine.

An implementation oddity is that we only allow uint32 sized allocations.
If your object is 2^32 sized or larger we can't allocate it. This is
because we imported the sized byte allocations from bytestore package.
These limitations made a kind of sense in bytestore (because we limit
the size the Reference was only 2 uint32 fields large). But it has no
upside here. This will be fixed in a future merge.