Add Race test for multi-type allocations
This test performs the multi-threaded tests of creating sharing
modifying and freeing allocations. In this new test we allocate objects
of a range of object sizes.

This new test does not assert _any_ result. It's _only_ value is
detecting data races. Asserting a result is very difficult with the
current test structure and seems very low value. Maybe a better test can
be made in the future which actually allows us to assert some kind of
result. I like these tests - they do find data races, I don't like these
tests they feel hard to read and difficult to modify.

We reduce the number of allocations performed in each test from 10_000
down to 1000. In practice the lower limit still finds races and the
tests complete faster.