Adding test for stats for each size class
These tests are a bit fragile, as we have to _carefully_ only allocate
objects of each size class only once. Because we track the number of
slabs allocated as well as raw/reused allocations asserting the correct
metrics quickly becomes difficult when we exercise the same size class
multiple times.