Allocate a1 byte for zero-sized type allocations
This change is a response to an intermittent failure case with zero
sized types.  The problem occurs because when we get the data portion of
the Reference we add the size of the meta-data to the pointer and treat
that as pointing to the data-portion of the allocation. In the case of
zero sized types this data-portion does not exist. In most cases the
data-portion will now point to the beginning of the next allocation's
meta-data, but for the last allocation in a slab it will point just past
the allocated memory region. If we get lucky the address just after the
allocated region is another allocated region, and everything works. If
we are unlucky the address just after the allocated region is invalid
memory and we segfault.

Because we cannot rely on being lucky, we now allocate a single byte to
ensure we can safely point to valid memory even though the zero sized
type won't use it to read/write.

There's probably a clever optimization for this. But this simple
approach will do for now.