Make the fuzz tests write the entire slice
Previously we only wrote up to 16 bytes into the allocation. This means
we do perform a write but for objects larger than 16 bytes most of the
allocated memory is unchanged.

We now write a single byte value across the entire allocation. This
makes the tests dramatically slower to run, but does ensure that we are
detecting any kind of data corruption across the whole of the allocation
region.

No test failures uncovered.