Split allocation and metadata in pointerstore
This reduces the allocation slab size for large allocations
dramatically. Because previously we would establish the memory needed
for the object (rounded up to a power of 2) and then add the metadata to
that and round the total up to a power of 2. With larger objects,
especially 1 object per slab objects, this basically doubles the size of
the memory allocated and wastes close to 50%.

Now we allocate the object memory and the metadata memory separately.
They are both still a power of 2 in size, but in the case of large
allocations take up much less memory.

Additionally we can now guarantee the alignment of objects, because they
are not pushed back by the metadata at the beginning of each allocation
slot. The details and guarantees of allocation alignment have not been
thought through deeply yet. So the thinking on this may change in the
future.

This has a real downside as each Reference now contains two pointers,
one to the object and one to the metadata. So the size of References has
now doubled. It's not clear at this stage how big of a problem this is,
and there may be a work-around for it in the future.