Mask data pointer value in pointerstore.Reference
We added a test to _explicitly_ pick up when the generation tag has not
been stripped out of the pointers in a Reference. This is required
because some platforms actually work fine with garbage data in the high
bits of an address, they just seem to ignore them. But this also means
that tests can run fine on one platform, but fail on another. So we use
a very explicit test to avoid that happening.

Also clean up the name of a linked_list test.