Adding objectstore.RefStr type
This reference type is used to allocate and manage strings inside an
objectstore.Store.

We include three alloction management functions for this.

	AllocFromStr(s *Store, str string) (RefStr, string)

Which allocates a string inside a *Store. The value of this string will
be copied from the `str` argument.

	AllocFromBytes(s *Store, bytes []byte) (RefStr, string)

Which allocates a string inside a *Store. The value of this string will
be copied from the `bytes` argument.

	FreeStr(s *Store, r RefStr)

Which frees an allocated string inside a *Store.

We also add a reasonable number of unit tests for this new
functionality.