Adding objectstore.RefSlice
This reference type is used to allocate and manage slices inside an
objectstore.Store.

We include three alloction management functions for this.

        AllocSlice[T any](s *Store, length, capacity int) (RefSlice[T], []T)

Which allocates a slice inside a *Store. The value pointed to by this
reference will be a []T with len and cap equal to length and capacity.

        FreeSlice[T any](s *Store, r RefSlice[T])

Which frees an allocated slice inside a *Store.

We also add a reasonable number of unit tests for this new
functionality.