Fixing bug in Slice allocation
The bug was that we choose the pointer-store based on the
typeSize*capacity of the slice. This is good. When we free we chose the
pointer-store using just capacity.

This is an easy bug to fix, but the more important part is finding the
test which would have caught it, so we can catch it if it's reintroduced
in the future.

The stats tests would catch this, as the stats for allocation would be
accumulated in one pointer-store while the free would be in another. So
we added the stats tests for both slices and strings.