Replacing bytestore.Store with objectstore.Store
The parcel_server can now use an instance of objectstore.Store to
allocate []byte to store in the quadtree.

In the future we might combine all the objectstore.Store's used here
into one. But we don't need to do that right now. Just eliminate the use
of the bytestore package here.