Adding Append() function for slices
This will take a slice-reference of type T, and append a value of type T
to it. The reference passed into the function is no longer valid after
the function returns. This is because if we required more space to store
the new value in the slice then we will free the original allocation and
copy the original slice contents across to a newly allocated slice.

This is very similar to standard Go append behaviour, but critically Go
doesn't invalidate your old slices. We do.

This implementation is missing a potentially significant performance
improvement. If the capacity of the slice is less than the total size of
the allocation slot it occupies we can just extend the capacity in
place. This avoids the alloc/free/copy work that would be done without
this optimisation.