Add AppendSlice() function
This function

func AppendSlice[T any](s *Store, into RefSlice[T], fromSlice []T) RefSlice[T]

Takes a slice-reference and returns a new slice-reference will the
elements of the fromSlice argument appended.

The into slice-reference must not be used after this function returns,
only the returned reference should be used. This is because the into
reference may have been freed if it didn't have enough capacity to
accommodate the new elements.

We add tests for this new function. And also switch all the slices in
the tests over to []byte. These smaller sized slices allow the tests to
exercise a greater range of allocation size classes. It also reduces
allocation size overall and speeds up the tests.