Combine AppendSlice tests into single test
Previously we had separate test cases for when the slice had enough
capacity to append an element without re-allocating and when it didn't.
These tests became longer and more complex as I thought about them
harder. But they also remained both almost exactly the same. So we
combined them into a single test with just one if statement for the
different capacity cases.