Always invalidate slice references on append
For both r.Append() and r.AppendSlice() we now invalidate the original
reference (r). This means that after the method returns calls to
r.Value() or FreeSlice(_, r) will panic. In general I doubt there's any
useful thing you can do with the original reference. It's dead.

As part of this change we added a small optimisation. When growing a
slice to allow elements to be appended it's possible that the slice does
not have enough capacity for the new elements, but the size of the
allocation does offer enough capacity for that growth. In this case we
reallocate the slice in its current location and grow the capacity.