Add ConcatSlices[T](*Store, ...[]T) function
This function takes an arbitrary number of slices and allocates a new
slice reference pointing to a slice whose elements are the elements of
the slices all allocated together.

This function is likely useful in general. But more speficically it is
needed if someone wants to append to a slice-reference without
invalidating the original slice-reference. The ConcatSlices function
does not invalidate any existing references, but it also must _always_
fully allocate new memory to hold the new slice-reference's data.