Adding generic resizeAndInvalidate() function
This function takes an existing pointerstore.RefPointer and re-allocates
it for a new size. After the function returns the existing
pointer-reference is invalidated. A new valid pointer-reference is
returned.

There are two cases.

If the new size is the same or smaller than the old size. The allocation slot is
left in place and the pointer-reference is realloc()'ed to invalidate
the old reference.

If the new size is larger than the old size, but the new size still fits
into the same allocation slot. The allocation slot is left in place and
the pointer-reference is realloc()'ed.

If the new size is larger than the old size, and the new size needs a
larger allocation slot. We allocate a new pointer-reference copy the
data from the old allocation across to the new one and free the old
allocation.

This function replaces the ensureCapacity function which performed this
utility prevously. The difference is that the new function doesn't
understand the sizes of types. It is given just the raw total size of
the old and new allocation sizes. This allows us to use the same
function for slices and strings.