Allow 31 size slots on 32 bit architecture
Previously we initialised 48 sized slots. Allowing allocations of up to
2^48 bytes.

On 32 bit architectures that was unnecessarily large. In particular we
currently limit the size of allocations by checking against the number
of sized allocation slots available. If your object is too large it's
slot index will be past the end of the available slots. This will
produce a panic. But since we over-provision slots for 32 bit
architectures there is no effective size limit.

We have chosen 2^31 as the size limit on 32 bit architectures. This was
chosen because slices have capacity and length in 'int' for vanilla Go.
So this seems like a very reasonable cutoff. I don't think people will
want to allocate more than 2^31 on 32 bit architectures in general
(maybe I'm wrong about that though).

Once we have this limit in place we can then use variable sized integer
values to represent object sizes in the codebase. We can drop the uint64
values. Maybe everything can just be 'int', or maybe it makes more sense
to make all sizes 'uintptr'. Haven't decided yet.