Removing capitalised error messages
I _think_ I like this rule. It's intention is to make regex checking of
logs more reliable by just eliminating capital letters. It can _feel_
awkward when you are writing the error messages, though. Because they
are english sentences and punctuation and capitalisation _feels_ nice.

For now we keep it without capitals. I wonder if there are more
aggressive restrictions on errors/logs which would make regexes easier
to write.