Adding object_examples_test.go
This contains all of the examples that are in the comments in
object_store godoc comments. They are all updated to actually compile
and work. We also have adopted the policy of writing all variables out
as a 'var name type'. The hope is that, as this code is primarily
documentation, this should help first time users quickly clarify the
types they would expect to be actually using with each method.