objectstore package renamed to offheap
This is shorter. It also avoids stuttering with types like
objectstore.Store or ojectstore.RefObject.

But most importantly it actually describes, to some extent, what this
package does. It store allocations off the main go heap.

I _might_ be a little java-esk. As this is what manual memory solutions
are generally called in java-land. Hopefully no one will hold that
against us.