First version of intern package
The intention is to have a package which interns []byte, int64 and
float64 as string values.

The interned string values are stored using the offheap package. This
means that we pay no garbage collection cost for the strings which are
interned but not currently in use.

I'll add a quick explanation here of the distinction between interned
string which are in use from those not in use. A string which is
interned but not in use is allocated on an offheap heap. There exists a
RefString reference pointing to it, but it is invisible to the garbage
collector. For these strings we pay no cost during garbage collection
cycles. When someone actually uses the interned string, however, they
are given a normal 'string' type. This 'string' contains a normal go
pointer (which points into the off-heap memory) but _is_ visible to the
garbage collector. If the users of the intern package intern strings and
then store all interned strings (normal string types) in some other
datastructure they will pay the usual garbage collection cost. The case
where the intern package will provide the most benefit is where the
string values taken from the intern cache are used for a short period of
time. A common example of this would be string values used when
servicing an http request.

A specific limitation of this intern package is that interned values
live forever. So each intern cache has a max bytes limit. An intern
cache will intern strings until it reaches this limit. After that no
interning will occur. The limit can be large (if you have enough memory
available) with no garbage collection cost. But the intern cache will be
most effective when operating on a limited set of values which repeat
often.