Adding Reference type constraint interface
This is intended to be used in generic data structures which are
intended to be used with some kind of reference type.

The Reference type can be any one of RefString, RefSlice or RefObject.

An example would be a map pointing to a Reference type. The type
declaration can be written like

    type ReferenceMap[T any, R offheap.Reference[T]] struct {
            rMap map[int]R
    }

We can then create an instance of this type, for a RefSlice[byte]

    var refSliceMap *ReferenceMap[byte, offheap.RefSlice[byte]]
        = NewReferenceMap[byte, offheap.RefSlice[byte]]()

A declaration of this map type for RefString is slightly awkward,
because we have to provide a type-constraint T which isn't used

    var refStringMap *ReferenceMap[byte, offheap.RefString]
        = NewReferenceMap[byte, offheap.RefString]()

By convention we always set this T constraint to 'byte'. But as it is
unused it can be anything at all.

The example tests are well intentioned but very nearly useless. The
generated docs do not actually include the type declaration for the
ReferenceMap. This declaration is the most complicated and non-obvious
part of using the Reference constraint type. We need to improve these
example tests to actually provide useful examples of the Reference type
in action.