Sharding string interner
By default we create a shard per CPU being used by the Go runtime. There
is an alternative constructor which allows a specific number of shards
to be created.

But regardless of the number of shards requested or the number CPUs used
we always round up the number of shards to the nearest power of two.
This allows fast division by bitmasking to be used. It's not 100% clear
that the performance advantage of bitmasking matters here, but this is
my default habit. May change in the future.

We ensure that the total bytes interned limit is not exceeded by
creating an internController type which is shared between all shards.
Using this we can CAS in new byte counts. This will definitely be a
point of contention for interning strings at very high rates.

Currently I doubt interning strings at incrediblly high rates will be a
critical need. For example if we imagine a web-server interning common
strings like status and header keys/values then it is unlikely that we
will see high new string intern rates at levels which bottleneck on CAS.
However, there may be other uses which really do stress this code path.
We can likely improve this in the future.

The names of types is currently quite chaotic, and is probably the next
thing to tidy up.

No tests yet. That's definitely the thing after next.