Share single *offheap.Store between intern shards
We _could_ use a separate offheap store for each shard. This might come
with performance benefits of keeping these datastructures very separate.
But it would likely waste a lot of memory.

Having a store per shard was an oversight by me, and is fixed here. But
we may reconsider this in the future if some very pressing performance
concern made this choice look appealing.

I doubt it will be needed though.