Adding stats to StringInterner
The stats are accumulated in each shard and then gathered together and
summarised.

Looking over this code it does pop out to me that we could have a
separate lock for each of the intern map/stats pairs. This would allow
ints and floats and bytes to be interned concurrently with each other.
Seems like a sensible and straight-forward change to make.