Give each intern map/stats pair a lock
This allows us to lock independently for bytes, int and float interning
operations on the same shard. Potentially increasing concurrency for
some workloads.

Previously I was considering using a read/write lock instead of the
basic lock used here. As the code is right now a full lock is needed
because even reading paths perform writes to stats fields even on pure
read code paths.

We could avoid this need by making the stat fields into Atomic.Int64
fields. This would allow read locks to write to them concurrently. But
the performance impact of this change is not easy to guess in advance. I
can imagine that the cost of atomic adds could outweigh the advantages
of increased concurrency of a read/write lock. Making this choice would
require convincing benchmarks. So we will keep the simple implementation
hwe have now until those benchmarkds exist.