Unit tests for intern package added
These tests are very basic. Testing out the basic functionality of the
bytes part of the interner.

There are two interesting observations of these tests.

1: The three test files are copy and pasted from the first test file
   written. They are basically almost exactly the same as each other
between the three byte/int/float methods. Is there a way to genericise
this and make them actually the same? Would this change break the
clarity of the implementation or tests?

2: There isn't much very clear difference in behaviour between a value
   being interned or not interned, the string representation of it is
returned regardless. The tests then rely on the stats counters to
determine whether the expected behaviour has occurred. This is _ok_, but
if there are bugs in the stats counters they could easily shadow bugs in
the interner. This seems quite bad.

One possible alternative to stats based verification would be to hold
onto a copy of an interned string, and then verify that when we retrieve
that interned value again it is actually pointing to the same location
i.e. &str1[0] == &str2[0].