Assert that interned strings are identical
Previously we were asserting that the strings returned had the value we
expected _and_ asserting that the stats had the values we expected. What
this testing approach missed is confirming that interned values are
actually re-using a single memory allocation. This is actually the
critical behaviour that the intern package exists for.