Adding some documentation to intern package
While going over the exposed interface of this package it feels clear
that the approach of supporting interned string conversion for []byte,
flat64 and int64 will be insufficient for most real world use-cases.

The first limitation observed here is that there are more types, many
more, which will generate strings and which could potentially benefit
from interning. If your type isn't one of the three chosen this whole
package is likely worthless to you. A good example of this would be
time.Time.

The second limitation is that we have fixed the format which ints and
floats are converted to strings. These formats are probably popular, but
people do use other formats, especially for floats, and if you wanted
something else this package is now completely useless to you.

It seems that we will need to build a more generic way for users
themselves to build these converters. Then instead of this interface
which allows you to convert three different types you will build a
single interner for a specific type.

Should be worth a try to see if this is feasible.