maxLen and maxBytes are unrestricted at <= 0
For maxLen we allow strings of any length to be interned if maxLen <= 0.
Previously this was -1.

For maxBytes we allow any amount of accumulated bytes to be interned if
maxBytes <= 0. Previously this was -1.

For maxBytes we now count the accumulated interned bytes even when there
is no limit. This allows accurate accounting of the total number of
interned bytes in the case where there is no limit.